Implemented purgetohere command
